+++ src/__generated__/ShowTestsQuery.graphql.ts
@@ -332,8 +332,8 @@ fragment Shows_show on Show {
   nearbyShows(first: 20) {
     edges {
       node {
-        ...ShowItem_show
         __id
+        ...ShowItem_show
       }
     }
   }
@@ -758,7 +758,7 @@ return {
   "kind": "Request",
   "operationKind": "query",
   "name": "ShowTestsQuery",
-  "id": "35f061a339070d46199bd6b7b9ad2765",
+  "id": "4534ea8e115e15baa7f77f2a652902a6",
   "text": null,
   "metadata": {},
   "fragment": {
@@ -1078,6 +1078,7 @@ return {
                     "concreteType": "Show",
                     "plural": false,
                     "selections": [
+                      v1,
                       v2,
                       v6,
                       v3,
@@ -1115,8 +1116,7 @@ return {
                             ]
                           }
                         ]
-                      },
-                      v1
+                      }
                     ]
                   }
                 ]