+++ src/lib/Scenes/Favorites/Components/Shows/index.tsx
@@ -89,6 +89,7 @@ export default createPaginationContainer(
             }
             edges {
               node {
+                __id
                 ...ShowItemRow_show
               }
             }